add script for calculating MAD and ACF
Rob Scharpf recommended calculating the median absolute deviation and
lag 10 autocorrelations to assess the quality of the arrays. This code
was used to evaluate the quality of the control samples and the values
were good according to Rob. Note that the VanillaICE pacakge also has
functions to calculate these values, which are very similar.