add alternate crlmm code
Instead of using the crlmm function to process the array data, we will
load in the FinalResult file, apply the gcCorrect function from the
ArrayTV package and use VanillaICE to call CNVs. This is the workflow
recommended by Rob Scharpf.

The crlmm package can't be used because the required CDF file is
missing.